refactor code

+++ crates/rwasm/machine/src/shape/mod.rs
@@ -190,7 +190,7 @@ impl<F: PrimeField32> CoreShapeConfig<F> {
                     );
                     for allowed_log2_height in allowed_log2_heights {
                         let allowed_height = 1 << allowed_log2_height;
-                        if height <= allowed_height {
+                        if height as u32 <= allowed_height as u32 {
                             for shape in self.get_precompile_shapes(
                                 air,
                                 *memory_events_per_row,
@@ -437,7 +437,7 @@ impl<F: PrimeField32> CoreShapeConfig<F> {
 
     pub fn estimate_lde_size(&self, shape: &Shape<RwasmAirId>) -> usize {
         // println!("shape:{:?},",shape);
-        shape.iter().map(|(air, height)| self.costs[air] * (1 << height)).sum()
+        shape.iter().map(|(air, height)| self.costs[air] * (1 << height) as usize).sum()
     }
 
     // TODO: cleanup..
